Moved Move_new_kanji_notes_to_learning_kanji_deck unit test to KanjiServiceManagementTests.
