+++ AnkiJapaneseFlashcardManagerTest/ApplicationLayer/Services/Managements/KanjiServiceManagementTests.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using AnkiJapaneseFlashcardManager.DomainLayer.Entities;
+using AnkiSentenceCardBuilder.Controllers;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -8,6 +10,43 @@ namespace AnkiJapaneseFlashcardManagerTests.ApplicationLayer.Services.Management
 {
 	public class KanjiServiceManagementTests
 	{
+		//TODO: MOVE TO GLOBAL VARIABLE
+		string _anki2FolderPath = "./Resources/Anki2 Files/";
 
+		[Theory]
+		[InlineData("emptyLearningKanji_1ivl飲12ivl良monthsIvl食欠人newKanji_decks.anki2", new[] { 1548988102030, 1552619440878, 1559353225229, 1559353240186 }, 1670499333507, 1708110494009)]
+		public void Move_new_kanji_notes_to_learning_kanji_deck(string anki2File, long[] expectedNoteIdsToMove, long expectedFromDeckId, long expectedToDeckId)
+		{
+			//Arrange
+			string originalInputFilePath = _anki2FolderPath + anki2File;
+			string tempInputFilePath = _anki2FolderPath + "temp_" + anki2File;
+			File.Copy(originalInputFilePath, tempInputFilePath, true);//Copy the input file to prevent changes between unit tests
+			Anki2Controller anki2Controller = new Anki2Controller(tempInputFilePath);
+			List<Card> allOriginalCards = anki2Controller.GetTable<Card>();
+
+			//Act
+			bool movedNotes = anki2Controller.MoveNewKanjiToLearningKanji();
+
+			//Get Assert Values
+			//Get changed cards
+			var allCardsAfterFunction = anki2Controller.GetTable<Card>();
+			var changedCards = allOriginalCards
+				.Join(allCardsAfterFunction,
+					originalCard => originalCard.Id,
+					updatedCard => updatedCard.Id,
+					(originalCard, updatedCard) => new { OriginalCard = originalCard, UpdatedCard = updatedCard })
+				.Where(pair => !pair.OriginalCard.Equals(pair.UpdatedCard))
+				.ToList();
+			//Assert
+			movedNotes.Should().BeTrue();//Function completed successfully
+			changedCards.Select(p => p.OriginalCard.DeckId).Should().AllBeEquivalentTo(expectedFromDeckId);//Original deck id should match
+			changedCards.Select(p => p.OriginalCard.NoteId).Distinct().Should().BeEquivalentTo(expectedNoteIdsToMove);//Original notes should match
+			changedCards.Select(p => p.UpdatedCard.DeckId).Should().AllBeEquivalentTo(expectedToDeckId);//Updated deck id should match
+			changedCards.Select(p => p.UpdatedCard.NoteId).Distinct().Should().BeEquivalentTo(expectedNoteIdsToMove);//Updated notes should match
+
+			//Cleanup
+			anki2Controller.Dispose();
+			File.Delete(tempInputFilePath);
+		}
 	}
 }
